pyproject.toml
[project]
name = "voiceagent"
version = "0.1.0"
description = "AURA Voice Agent - AI Mental Health Counselor powered by LiveKit"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "livekit-agents>=0.12.0",
    "livekit-plugins-openai>=0.10.0",
    "livekit-plugins-cartesia>=0.4.0",
    "livekit-plugins-silero>=0.7.0",
    "livekit-plugins-deepgram>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
voiceagent = "voiceagent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voiceagent"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
strict = true
